Moves make_pipeline_definition into node method.
