+++ podpac/core/pipeline/__init__.py
@@ -1,3 +1,2 @@
 from podpac.core.pipeline.pipeline import Pipeline
 from podpac.core.pipeline.util import PipelineError
-from podpac.core.pipeline.util import make_pipeline_definition
