+++ podpac/core/pipeline/test/test_pipeline_imports.py
@@ -2,4 +2,3 @@
 def test_import():
     from podpac.core.pipeline import Pipeline
     from podpac.core.pipeline import PipelineError
-    from podpac.core.pipeline import make_pipeline_definition
