+++ podpac/pipeline.py
@@ -3,9 +3,5 @@ Pipeline Public Module
 """
 
 from podpac.core.pipeline import Pipeline, PipelineError
-from podpac.core.pipeline.util import (
-    parse_pipeline_definition, make_pipeline_definition,
-)
-from podpac.core.pipeline.output import (
-    Output, NoOutput, FileOutput, FTPOutput, S3Output, ImageOutput,
-)
+from podpac.core.pipeline.util import parse_pipeline_definition
+from podpac.core.pipeline.output import Output, NoOutput, FileOutput, FTPOutput, S3Output, ImageOutput
